Add an option to fully optimize the variables.
Currently, if a user supplies variables on the command line, it will only
optimize the variables after that (allowing the user to generate a new route for
a run in progress). This way, the user can input a route and use it as seed data
for a complete optimization run.

Signed-off-by: Jason Lynch <jason@calindora.com>