Normalize variable values after optimization.
This will ensure that regardless of what magic the optimization algorithm
performs, the variable values will be returned to sane values. The normalization
function will minimize all values to the lowest value that keeps the total
duration of the route the same. This will ensure that steps above the maximum
for the area are discarded and that flag variables are set to their maximum
value rather than some arbitrarily high value.

Signed-off-by: Jason Lynch <jason@calindora.com>