The pair and bb algorithms should now prefer steps later than earlier.
Because no distinction is made in the variable sets between choices and
steps, this will probably miss some corner cases. The whole variable
structure is kind of silly, anyway.

A side effect of this change is that the other algorithms will no longer
prefer a version of the route that takes the same number of frames but
uses fewer variables. It is therefore advised to follow up any ILS use
with the simplified pair algorithm. (-a ils+pair -z with appropriate
configuration parameters should suffice).

Signed-off-by: Jason Lynch <jason@calindora.com>