Switch to a CMake-based build system.
I'm not particularly happy with how this is set up, and if tup ever
gains the ability to generate the clang JSON database, I will most
likely switch back. That said, it works fine for now. It is still
required to build using the build script, unless I can figure out a way
to do equivalent things directly using CMake.

Signed-off-by: Jason Lynch <jason@calindora.com>