Handle all times internally using integer milliframes.
This allows us to both retain all the existing precision we had, as well
as eliminate any problems with comparing doubles or testing them for
equality. Some of the way things are handled is temporary, but I'm
trying to do this in small sections that still compile.

Note that this commit changes the meaning of the FRAMES variable in the
route output files from frames to milliframes and will therefore not
compare direclty with previously generated routes. This isn't a huge
issue, because all of that should be refactored entirely by the next
release.

Signed-off-by: Jason Lynch <jason@calindora.com>