Add initial support for explicitly-named variables.
This commit does what it claims to do, but it also disables all
optimizatino features, disables automatic route output and disables
automatic route loading. As it turns out, adding proper support for
explicitly-named variables reaches deep into the program, and it would
not be worth the effort to fix a lot of features that are about to be
gutted anyway.

In short, if you want anything resembling useful results at this time,
stick to the commit before this one.

Signed-off-by: Jason Lynch <jason@calindora.com>