Begin to implement the optimal solver.
Early tests look good, but this has not been tested thoroughly yet, and
bugs may very well be found in time. This commit represents a drastic
reduction in functionality. Routes will be optimized, but the output is
very limited.

This is essentially a WIP commit that represents the first marginally
usable state since I began the modifications.

Signed-off-by: Jason Lynch <jason@calindora.com>