Handle SEARCH/WAIT differently.
The previous version would simply keep taking extra steps in the final
search room until the targets were found. These implicit steps would not
be recorded in the variables which caused much consternation.

The next solution was to simply declare a solution that didn't find the
targets infeasible and return the maximum possible value. Unfortunately,
this leads to a base unoptimized route being invalid, and giving us
nothing useful to compare against.

The new solution returns to taking extra steps, but does so in a way
that they are recorded properly in the variables.

Signed-off-by: Jason Lynch <jason@calindora.com>