Eliminate unused options and reactive command-line variable parsing.
A neat new feature of the variable parsing is the ability to fully
constrain variables' ranges instead of merely specifying a value. I'm
not sure it has tons of practical value, but it might be fun for playing
around with.

Note that variables are named as valid hexadecimal numbers. The code is
more lenient, but I have standardized on using a single letter to help
me notate the type of variable, 4 numbers to encode the map index, and 2
additional numbers to provide an index in case a single map has multiple
variables.

Finally, in the event of an encounter search, maximum constraints will
be ignored as necessary to find the target encounter.

Signed-off-by: Jason Lynch <jason@calindora.com>