Rename the project to Rosa and update the README.
The original name, spoony, was just a pointless name and since the
character of the project has changed so much with this new version, I am
renaming it to Rosa, which was the name of my experimental rewrite. I
ultimately decided to modify the original codebase instead. Rosa is a
recursive acronym for "Rosa Optimizes Steps Automatically".

Signed-off-by: Jason Lynch <jason@calindora.com>