Update the no64-rosa routes.
This modifies the primary no64-rosa route to skip the healing pot in the
Castle of Dwarves. The -safe route variant has been added which keeps
the healing pot and additionally adds the two elixirs in Cave Eblana, to
enable the safe usage of the Rosa+Rydia safety strats for Zeromus.
Finally, the -marathon variant adds a visit to the Hummingway shop (to
purchase elixirs) and makes several save points mandatory. It also adds
numerous reminders about these save points and other potential places
for safety saves.

Signed-off-by: Jason Lynch <jason@calindora.com>