Fix a memory issue with the search expression parser.
The library takes the string as a string_view. If that string ever goes
out of scope (if the instruction gets copied, for instance), the
expression AST will no longer have valid data. This is solved by putting
that string behind a shared_ptr, so that only one copy of the string
will ever exist, even if the instruction gets copied multiple times.

Signed-off-by: Jason Lynch <jason@calindora.com>