Add the ability to add a penalty to the first encounter on a given segment.
Experimental data has determined that the first encounter on a map in
Final Fantasy IV is generally longer by a small number of frames. This
difference is not consistent throughout the game, and is likely related
to the exact conditions at the time. The existing routes do not
currently make use of this functionality.

Signed-off-by: Jason Lynch <jason@calindora.com>