Remove the potential Hummingway trip before FuSoYa in no64-rosa-marathon.
As it turns out, none of the seeds wanted to use it anyway. Presumably,
as any potential steps that would be useful there can almost certainly
be done at a slightly lower cost in the Sealed Cave. Aside from that,
the timing was not accurate, as it did not include the time spent flying
to and landing at the cave.

Signed-off-by: Jason Lynch <jason@calindora.com>