Add the initial version of the glitchless route.
This may very well be modified frequently in the future, but at this
point, it's good enough to commit an initial version. The most obvious
changes are to consolidate some of the overly expanded areas and to
remove irrelevant segments. The existing route may be infeasible to
compute to even as low as 128 steps per segment.

Signed-off-by: Jason Lynch <jason@calindora.com>